Use port as like HTTP request
In JS it was returning new data everytime when you request store via port but it caused a problem which comes when you want sequencial job via port and then fetch final data because whenever new data comes from port the subscription works based on current model which is immediately changed before final data comes.

When a data need to be created or modified, Use port as like HTTP request to store and fetch data

When a data need to be created or modified, it needs to be done by sending value via port to JS to updates localStorage data and update current model as like 'FetchingSites' to fetch new data from JS.

In previeous JS code, it was returning new data everytime after when you request store via port but it caused a problem which comes when you want sequencial job via port and then fetch final data.

Because whenever new data comes from port, the subscription works based on current model which is immediately changed before final data comes in.

So the JS was changed to not send new data automatically, it sends only when fetch function called.

As Cmd.batch is not for sequencial commands, that's where "getResultAfter" function comes in to be used like "Result" which explicitly invoke "Msg" after port jobs you want.
send value via port to JS to updates localStorage data and update model as like 'FetchingSites' to fetch new data from JS.

Cmd.batch is not for sequencial commands, that's where "getResultAfter" function comes in to be used like "Result" which explicitly invoke "Msg" after port jobs you want.