Improve pmList viewer, update DAO
The MDView actionBar view updated to be sticked on top.

The PlayerManager List view updated as a table.

The full spec of DAO for PlayerManager and Player has been fulfilled.

Elm JSON decoder is not supporting mapping with over 8, so an options field added to contain rest fields to Player DAO.