Refactor Model and Msg naming which was ambiguous
Especially like FetchingSites and FetchingSiteList and others.

The Detail page is intruduced to show the PlayerManager's detail information
and list of players. (not implemented fully yet)

Some functions for port fixed to support the changes.

The pmModel merged into Site record to help understanding the
mental model of the application.

It took about 5 hours... damn..